refactor(notifications): unified approval card data resolver
Resolve a notification's contextPayload into a typed ApprovalCardData
discriminated union (access_request | tool_approval) in one place, then
render it to Surface [ui_surface, text] blocks via the shared
buildApprovalCardBlocks. Both approval types now flow through a single
resolveApprovalCardData dispatcher + renderApprovalCardData renderer
instead of each copy module independently parsing contextPayload.

- New approval-card-data.ts owns the union, dispatcher, renderer, and the
  two Surface seed-content builders.
- access-request-copy.ts keeps its parsing / contract-text / warning
  primitives; its seed builder moves to approval-card-data.ts.
- tool-approval-copy.ts is fully absorbed into approval-card-data.ts.
- copy-composer.ts and decision-engine.ts import both seed builders from
  the new module.

Seed-content output is byte-identical: 34 seed-content tests + 19
copy-composer tests pass, tsc --noEmit and eslint clean.

Realizes the intent of #34916 (unified approval card dispatcher) on main's
current schema-first notification pipeline, without reverting the Zod
migration or the pre-resolved accessRequestContext path.
